## Summary

- Upgraded all Go dependencies to latest versions (`go get -u ./...` + `go mod tidy`)
- Pinned `slog-fiber` back to v1.18.0 — v1.22.2 switched to the Fiber v3 handler signature and is incompatible with Fiber v2
- Fixed `uint` → `uint64` cast errors introduced by gorm v1.31 tightening implicit numeric conversions (6 sites across services and types)

+++ pkg/services/activities.service.go
@@ -368,7 +368,7 @@ func (s *ActivityService) getUserMap() (map[uint64]types.User, error) {
 
 	userMap := make(map[uint64]types.User)
 	for _, user := range users {
-		userMap[user.ID] = *types.UserFromDal(&user)
+		userMap[uint64(user.ID)] = *types.UserFromDal(&user)
 	}
 	return userMap, nil
 }
