+++ pkg/services/auth.service.go
@@ -54,7 +54,7 @@ func (s *AuthService) Login(req LoginRequest) (*LoginResponse, error) {
 	}
 
 	return &LoginResponse{
-		UserID:   u.ID,
+		UserID:   uint64(u.ID),
 		Email:    u.Email,
 		LoggedIn: true,
 	}, nil
@@ -114,12 +114,12 @@ func (s *AuthService) GetUsers(currentUserID uint64) (*GetUsersResponse, error) 
 	res := make([]*types.UserForLogin, 0, len(users))
 	for _, v := range users {
 		user := v
-		if v.ID == currentUserID {
+		if uint64(v.ID) == currentUserID {
 			continue // Skip the current user
 		}
 
 		res = append(res, &types.UserForLogin{
-			ID:    user.ID,
+			ID:    uint64(user.ID),
 			Name:  user.Name,
 			Email: user.Email,
 		})
